+++ R/stats.R
@@ -65,12 +65,11 @@ stages_stats <- function(e){
 }
 
 #' Compute TST90, the percentage of time during sleep with an oxygen saturation below 90%.
-#'
+#' @description Compute TST90, the percentage of time during sleep with an oxygen saturation below 90%.
 #' @param spo2_signal The SpO2 signal vector.
 #' @param sRate The SpO2 signal vector sRate.
 #' @param startTime The SpO2 signal start time.
 #' @param hypnogram Events dataframe containing hypnogram.
-#' 
 #' @return TST90 value.
 #' @export
 tst90 <- function(