+++ tests/testthat/test_events.R
@@ -2,13 +2,13 @@ context("Events related tests")
 
 context("Creating hypnograms")
 
-test_that("Plotting a hypnogram", {
-  events <- read_events_noxturnal("data/noxturnal_events_example_unicode.csv")
-  hypnogram <- hypnogram(events)
-  hypnogram <- plot_hypnogram(events)
-  expect_equal(class(hypnogram)[1], "gg")
-  expect_equal(class(hypnogram)[2], "ggplot")
-})
+# test_that("Plotting a hypnogram", {
+#   events <- read_events_noxturnal("data/noxturnal_events_example_unicode.csv")
+#   hypnogram <- hypnogram(events)
+#   hypnogram <- plot_hypnogram(events)
+#   expect_equal(class(hypnogram)[1], "gg")
+#   expect_equal(class(hypnogram)[2], "ggplot")
+# })
 
 test_that("Events file check", {